+++ chakracore-cxx/lib/Backend/IRViewer.cpp
@@ -482,8 +482,8 @@ void IRtoJSObjectBuilder::CreatePragmaInstruction(Js::ScriptContext *scriptConte
     int32 statementIndex = (int32)inst->m_statementIndex;
     LPCUTF8 sourceBegin = NULL;
     LPCUTF8 sourceEnd = NULL;
-    ULONG line = 0;
-    LONG col = 0;
+    uint32_t line = 0;
+    int32_t col = 0;
 
     Js::FunctionBody *fnBody = func->GetJnFunction()->GetFunctionBody();
     fnBody->GetStatementSourceInfo(statementIndex, &sourceBegin, &sourceEnd, &line, &col);
